requirements.txt
flask
langchain
langchain-experimental
langchain-openai
python-dotenv
gunicorn
werkzeug
pandas
numpy
tabulate
psycopg2-binary
